+++ src/commands/sentry/RemoveMuteCommand.js
@@ -37,6 +37,8 @@ class RemoveMuteCommand extends Command {
 			msg.reply("Invalid mute ID, make sure that you are using the number from their history.");
 		}
 		
-		msg.delete();
+		if (msg.channel.id !== process.env.STAFF_COMMANDS_CHANNEL) {
+			msg.delete();
+		}
 	}
 }
