+++ src/commands/sentry/UnmuteCommand.js
@@ -34,6 +34,8 @@ class UnmuteCommand extends Command {
 			msg.reply("That user isn't muted.");
 		}
 		
-		msg.delete();
+		if (msg.channel.id !== process.env.STAFF_COMMANDS_CHANNEL) {
+			msg.delete();
+		}
 	}
 }
